Generate docs from schema file
Previously, the docs were generated by sending an introspection
request to a running instance of the API Gateway. Instead of doing
this, we can use a file:// protocol rather than http:// and generate
the docs directly from a copy of the GraphQL schema.

Our tests in the API Gateway already ensure that the schema matches up
to the implementation of the resolvers, so there is no need to use a
running instance of the API Gateway component here.